UCA 14 new repertoire
From Ken:

uc130to140add.txt

That is the entire new 14.0 repertoire, extracted from
UnicodeData-14.0.0d9.txt (but reformatted to the shorter lines with
omitted irrelevant fields that I use for the sifter input). Omitted also
are the lines related to range change adjustments for CJK ideographs --
those don't impact the explicit input file, but are handled with
constants in the sifter code, instead.