UCA 14 delta 6
From Ken:

unidata-14.0.0d6.txt

uc130to140add.d6.txt

For this delta, I took care of a bunch of small script interpolations --
all of them except for the Arabic, Latin, and Katakana additions, each
of which has special issues. Altogether, this delta takes care of 64
more new characters. In more detail:

1. Balinese 1 character: 1B4C. This one got a synthetic decomposition to
<ja, adeg adeg, nya>, as per the suggested collation in the proposal
document. So it also resulted in one more line in the output decomps.txt
file.

2. Tagalog 3 characters: 170D, 171F, 1715. Intercalated as per the
proposal document.

3. Glagolitic 1 case pair: 2C5F/2C2F. No information on collation in the
proposal document, so instead of attempting an intercalation, the case
pair was just added at the end, along with some other variant letter forms.

4. Kaithi 1 character: 11062. Simple intercalation.

5. Brahmi 6 characters: 11070..11075. Intercalated per detailed
instructions in the proposal document.

6. Ahom 7 characters: 11740..11746. For this one I threw up my hands.
The new additions were to fill out Ahom for Pali, but the problem is
that the main Ahom letters already don't exactly follow a Sanskrit
ordering. The proposal document was silent on collation. So I ended up
slotting them in just using code point order at the end of the run of
the other Ahom consonants, following some other late-added characters in
that same location.

7. UCAS 16 characters: 11AB0..11ABF. The default order for the Syllabics
is basically hopeless, as the characters are intended for multiple
orthographies used across three very different language families. The
order in DUCET already just uses code point order for the syllabics, so
I just tacked on the new block at the end. Any particular Syllabics
collation requires language-specific tailoring.

8. Ethiopic 28 characters: 1E7E0..1E7FE. The proposal argued for
intercalating all these Gurage additions just ahead of each
corresponding legacy syllable, so that is what I did. The results are
not completely consistent with the prior way that some Sebatbeit
equivalents were intercalated, but are quite close. Note that the
proposal's table of equivalents had a few code point errors in it, but I
think I ended up with new additions in the right positions.

Output files are posted in the same directory:

allkeys-14.0.0d6.txt

decomps-14.0.0d6.txt

This allkeys.txt for this delta is no longer diffable with prior output
files, because of all the primary key additions this time.