UCA 14 delta 8
From Ken:

This takes care of adding Vithkuqi and Cypro-Minoan.
Vithkuqi and Cypro-Minoan, between them, account for 169 more of the
outstanding characters left after the prior delta.

unidata-14.0.0d8.txt

uc130to140add.d8.txt

1. Vithkuqi: 70 characters. Ordered after the other two Albanian scripts
in the table, Elbasan and Caucasian Albanian. This script is bicameral,
so I reordered the input entries to harmonize with the way this is done
for other bicameral scripts, and verified in the output that the case
tertiary distinctions are all calculated correctly. (That is a check on
my case mapping implementation.)

2. Cypro-Minoan: 99 characters. 97 of these are just gc=Lo signs, now
ordered after the Cypriot script. The other two are punctuation marks
that I grouped with other obscure Middle-Eastern ancient script divider
punctuation marks.

DUCET output file based on this input is also posted:

allkeys-14.0.0d8.txt