UCA 14 delta 12
From Ken:

And the remainder of the outstanding characters: Arabic additions. This
last delta accounts for the remaining 50 characters.

unidata-14.0.0d12.txt

Because there is no residue, the uc130to140add.txt file has been emptied
out, and there is no d12 delta for that file.

For Arabic, the characters were significantly underspecified for
collation in the proposals. In fact, one of the main proposal documents
for the Arabic additions, L2/19-306, said the following about collation:
"No collation information is available." Sum total. ;-)

I took whatever information there was, for a half dozen or so of the
additions, and just seat of the pants-ed the rest. The combining marks
for Quranic annotations were mostly just ignored. The big collection of
alefs with various attachments for Quranic annotation, were just equated
to alefs. The rest I just made guesses for, which in most instances
should be good enough for the default ordering. Note that the overall
Arabic ordering in DUCET just tries to put the skeletons in basic Arabic
order. Any meaningful ordering for any languages, including Arabic,
would require extensive tailoring, anyway. So close enough is probably
good enough for this newly added and somewhat obscure repertoire.

Output files:

allkeys-14.0.0d12.txt

decomps-14.0.0d12.txt