UCA 14 sifter code changes
From Ken:

The only file changed this time was unisift.c, and the changes there
were relative minor and localized.

Of course, I still have to verify that the CTT table generation for
14651 went o.k., and that might require some other fixes later on.

The minor impact on the code was helped greatly by making sure that I
disposed of all the new characters without requiring the addition of any
new secondary weights. Those secondary weight additions have the most
wide-ranging impact on the code, generally. By a bit of cleverness in
working all the new character records into the input file, I was able to
limit the impact of 14.0 only to the addition of new primary weights to
the table.