UCA 15 delta 2
From Ken:

This took care of 43 more characters which could be intercalated without
affecting any primary or secondary weights, so the allkeys.txt is still
easily diffable against the 14.0 version.

1. 10EFD..10EFF 3 Quranic annotation marks, all ignored.

2. 13440, 13447..13455 Egyptian mirroring mark and damage modifiers, all
ignored. Notionally, these are more annotation marks.

3. 0CF3, 11F00, 11F01, 11F03 Kannada and Kawi candrabindu, anusvara, or
visarga marks, all equated to the base weights for the equivalent
Devanagari marks.

4. Kawi digits and Nag Mundari digits, which weight equivalently to
0..9, as for all other sets of digits.

The next deltas will break diffability, as I will first take care of the
variables (symbols and punctuation), and then work on specifying the
primary and secondary order for all the alphabetic stuff that will
remain. The good news is that there are only 181 more characters to deal
with, and none of them appears to be particularly problematical.