UCA 15 delta 3
From Ken:

This delta added in all the punctuation, all the symbols, and the Kaktovik numerals.
The punctuation and symbols are variables,
but they impact the primary weight assignements, so allkeys.txt is no longer diffable.

1. The punctuation and symbols I just intercalated in whatever seemed the most likely locations.
The collection of Kawi miscellaneous punctuation could have gone in a couple of locations
in the input file, but I went with sticking it after some other groups of section signs.

2. Kaktovik numerals are comparable to Mayan numerals, so I treated them in parallel.
The zero through nine end up getting the same 0..9 weights as digits,
but the ten through nineteen just end up with symbol weights
along with all the rest of the miscellaneous numerals and fractions from various scripts.

109 more characters to deal with:
Kawi and Nag Mundari letters and marks, and then miscellaneous others.