UCA 15 delta 4
From Ken:

This takes care of the rest of the characters:

1. Kawi script. This was straightforward, except for two characters, the
JNYA and the REPHA. For now I just left the JNYA at the end of the
consonants, but in Balinese we have a precedent of expanding the cognate
form to JA + CONJOINER + NYA. I'll need to follow up with Norbert and
see what he thinks. The REPHA I just gave a primary weight right after
RA, but it might make sense to weight it as RA + KILLER.

2. Nag Mundari. Also straightforward. This script has four new combining
marks that I gave new secondary weights to.

3. Hieroglyphs one new sign, plus blanks and lost signs. All given
primary weights. I kept the blanks and lost signs at the end of the run
of hieroglyphs.

4. 6 more Latin letters for an obsolete Malayalam transliteration
scheme. These are the "WITH RAISED LEFT HOOK" characters. I intercalated
them right after the corresponding HOOK or RETROFLEX HOOK letters.

5. 3 more Khojki letters/signs, treated as specified by Anshuman.

6. Lao Yamakkan (0ECE). The proposal said it "must collate after virama"
but at face value that would imply giving it a primary weight, which
seemed inadvisable and inconsistent with the weighting of the cognate
character in Thai. I gave it a new secondary weight, parallel to the
Thai treatment.

7. In a small cleanup, I changed the secondary weight for the Phaistos
Disc diacritic to a generic diacritic below. It didn't deserve a
secondary weight all its own. So the net net for 15.0 is the addition of
4 new secondary weights.

These additions required some more small changes to the sifter code, and
in particular, the adjustments I always have to make in the ISO 14651
branch of the code to deal with any new secondaries. I won't deliver the
code quite yet, because I have yet to scan through the generated CTT
table, to see if I overlooked anything in that part of the generation.