UCA 15 delta 5
From Markus:

I notice that among the variants of R there is a group of "turned R",
but the group is interrupted by a couple of non-turned Rs, including the new one.
Wouldn't it make sense to group all of the turned Rs with their legs and hooks
after all of the regular, non-turned Rs with their legs and hooks?

From Ken:

I agree with you. I had missed the "TURNED" part of the name there when
I was scanning down looking for "WITH HOOK".
The actual "r with hook" is U+027D LATIN SMALL LETTER R WITH TAIL,
which has a "WITH TAIL" name for arbitrary historical reasons. ;-)

I moved the 1D89 entry down past 027B/02B5, and then put 1DF28 after 027D/2C64/107A8.

From Markus:

Thanks, that looks better.
I would actually expect to group all of the regular-orientation Rs with squiggles,
then groups of all of the reversed, turned, inverted Rs each with their squiggle variants...
but I don't think it's important.