UCA 15 sifter code changes
From Ken:

The only changes for 15.0 are:

unisift.c

A bunch of version and date updates;
updates for CJK ranges;
http: --> https: updates in a couple outputs for terms of use;
tweaks to the kana mapping tables for the two new small kana characters.

unisyms.c

Updates to the symbol tables for 5 new combining marks
(and for removing the explicit separate weight for the Phaistos Disc symbol mark);
updates of SMP block ranges to make sure all the appropriate
ranges of symbolic weights get output correctly.

cttfooter.txt

Updates for some CJK ranges in the static text appended at the bottom of
the CTT table output.