UCA 16.0 delta 2
From Ken:

For this delta, search the input for all the new lines that can be
intercalated into unidata.txt without affecting any primary weights
or introducing any new secondary weights.

1. Move the new uppercase for 0264 (A7CB) into unidata.txt right below
the entry for 0264. This just introduce the new uppercase, and does
not impact the primary weight sequence at all.

Verify by generating allkeys.txt and examining the diff.

2. Look for any new combining marks for Brahmic scripts that should be
equated with the existing weights for Devanagari candrabindu, anusvara,
and visarga. (This is a regular feature of DUCET now, to avoid the
unnecessary proliferation of secondary weights for these for each
new script that has them, because they are never mixed and matched
across scripts.) The clear candidate for 16.0 is Tulu-Tigalari. The
3 characters in question are:

113CA;TULU-TIGALARI SIGN CANDRA ANUNASIKA;Mc;0901;;;;;
113CC;TULU-TIGALARI SIGN ANUSVARA;Mc;0902;;;;;
113CD;TULU-TIGALARI SIGN VISARGA;Mc;0903;;;;;

Checking the latest proposal document (L2/23-031) verifies that
"CANDRA ANUNASIKA" is the Tulu-Tigalari candrabindu analog.
These three lines are copied into unidata.txt right below the
corresponding entries for the Grantha analogs (closest related
script, as well as sequential in code point order). The artificial
decompositions 0901, 0902, and 0903 are manually added to the
entries.

Verify by generating allkeys.txt and examining the diff.

3. Rinse and repeat for the next new Brahmic scipt with any of these
three characters: Gurung Khema:

1612D;GURUNG KHEMA SIGN ANUSVARA;Mn;0902;;;;;

The order of intercalation for these entries is not critical, because they
are just being equated to something else. But I put it after 11F03,
KAWI SIGN VISARGA, to keep it in code point order in the input file.
The artificial deocmposition 0902 is manually added to the entry.

Verify by generating allkeys.txt and examining the diff.

4. Note that the Kirat Rai ANUSVARA, TONPI (a bindu), and VISARGA should
*not* be equated to the Devanagari combining marks. Kirat Rai was
deliberately encoded more like an alphabet, even though it minimally
qualified as an abugida because of the inherent vowel and the
existence of a killer. It is best for Kirat Rai to just give these
three characters a primary weight in the code point order. So I
skip over them for this draft of unidata.txt. They will be processed
later when adding all the primary weights for Kirat Rai. So a no-op
at this point in the processing.

5. Note any obvious sets of compatibility decompositions that will
just result in more equivalences without impacting primary weights.
The obvious set for 16.0 are the outlined Latin capital letters in
the legacy computer symbols repertoire, 1CCD6..1CCEF. Move these
entries into unidata.txt just before the very similar set of
squared Latin capital letters, 1F130..1F149. Again, the exact placement
in unidata.txt doesn't matter, because these will all end up equated
to existing other weights, but putting 1CCD6..1CCEF in that location
makes the parallel collation treatment obvious and makes these easier
to track in the input file. No manual modification of the decomposition is
needed, as these all already have formal compatibility decompositions.

Verify by generating allkeys.txt and examining the diff.

At this point, the obvious candidates (other than digits) have been
taken care of. 31 down, 1146 to go. Note that to this point, not only
is the diff for unidata.txt easy to examine, but also the diff for
allkeys.txt is still well-formed and easy to interpret.

Archive this delta 2:

unidata-16.0.0d2.txt (1552253 bytes, 10/07/2023)