UCA 16.0 delta 4
From Ken:

After scanning the remaining new characters for possible candidates for
intercalation without impacting primary or secondary weights, I spied
the following candidates.

1. Nuktas

There was a stray nukta added, 11F5A KAWI SIGN NUKTA. Most of the nuktas
are also folded to a single secondary weight, so this one won't require a
new secondary weight:

11F5A;KAWI SIGN NUKTA;Mn;093C;;;;;

Note that an explicit decomposition to 093C is added here.

Verify by generating allkeys.txt and examining the diff.

2. Vedic tone marks

The pattern in DUCET for all Vedic tone marks is just to ignore them
completely for collation. There are a couple Vedic tone marks in 16.0
added for Tulu-Tigalari. These can be added to the ranges of ignored
Vedic accents in unidata.txt.

Verify by generating allkeys.txt and examining the diff.

3. Garay combining marks

Garay has four diacritic combining marks, 10D6A..10D6D, which do not
participate in any canonical equivalences. The combining 10D69 GARAY VOWEL SIGN E
is highly significant to the orthography, so must get a primary weight.

10D6A;GARAY CONSONANT GEMINATION MARK;Mn;;;;;;
10D6B;GARAY COMBINING DOT ABOVE;Mn;;;;;;
10D6C;GARAY COMBINING DOUBLE DOT ABOVE;Mn;;;;;;
10D6D;GARAY CONSONANT NASALIZATION MARK;Mn;;;;;;

All 4 of these are non-spacing marks above.

The proposal is unclear about the collation implications of the gemination mark,
but it seems safe to presume it should be given a secondary weight and not be
ignored.

The dot above and double dot above are effectively two nuktas. The dot above is
used on one letter to mark a native Garay ŋ, as opposed to a prenasalization.
The dot above and double dot above are used as diacritics on another letter to
indicate two borrowed Arabic sounds. That means the dot above and double dot
above need to be distinguished from each other. For these diacritic marked letters
there are no atomic encodings -- they can only be represented as sequences.
The collation information in the proposal indicates that the sequences (e.g.,
the two sequences representing Arabic sounds: <10D76, 10D6B>, <10D76, 10D6C>)
should get primary weights. DUCET does this kind of thing for atomic characters
which have canonical decompositions, but it does *not* do it for arbitrary
sequences, since there is no "target" in the encoding to assign the primary
weight for the sequence to. Conceivably, the sifter apparatus could be extended
to allow for weighting these ghost primaries, but not now for the 16.0 draft.

The nasalization mark (10D6D) isn't clearly exemplified, and the proposal
claims it is ignored for collation.

The gemination mark can occur over the GARAY VOWEL SIGN E and also over the
COMBINING DOT ABOVE. At a minimum, its secondary weight should be distinguished
from the COMBINING DOT ABOVE.

There are other complications for the Garay collation which mean it cannot
be fully handled by a default ordering in DUCET anyway.

The net conclusions I draw for now are that:

a. 10D69 GARAY VOWEL SIGN E needs a primary weight.
b. 10D6A GARAY CONSONANT GEMINATION MARK should get a script-specific secondary weight.
c. 10D6B GARAY COMBINING DOT ABOVE can be weighted as generic above [0033]
d. 10D6C GARAY COMBINING DOUBLE DOT ABOVE can be weighted as generic nukta [00C2]
e. 10D6D GARAY COMBINING NASALIZATION MARk can be weighted as generic above [0033]

a) and b) will be dealt with later when doing all the primary weighting for Garay.
c), d) and e) won't impact primary or secondary weights already in DUCET, so I
am dealing with them now, and bleeding them out of the set of the Garay characters
to be weighted later. The relevant hacks for them are:

10D6B;GARAY COMBINING DOT ABOVE;Mn;F8F5;;;;;
10D6C;GARAY COMBINING DOUBLE DOT ABOVE;Mn;093C;;;;;
10D6D;GARAY CONSONANT NASALIZATION MARK;Mn;F8F5;;;;;

to equate them to a generic above secondary weight (F8F5) and to the generic
nukta (093C). These entries are added to unidata.txt in the section dealing
with secondary weighting, just ahead of the section discussing Adlam secondaries,
with notes explaining the generic weight decisions.

Verify by generating allkeys.txt and examining the diff.

This concludes all the apparent candidates that don't require new primary or
secondary weights -- although there are possibly a couple other gc=Mn that
might qualify. I'm deferring those to further analysis on a per-script basis.
6 more down, 1060 to go.

Archive this delta 4:

unidata-16.0.0d4.txt (1556145 bytes, 10/07/2023)