UCA 16.0 delta 5
From Ken:

I'm now at the point where further updates to unidata.txt will render the
output file (allkeys.txt) effectively undiffable. I've never felt it was
worth the effort to try to write custom tooling that would keep track of
all the relative differences in weights and report changes on those differences.
I suppose that is an opportunity for somebody who feels ambitious.

1. New Latin case pair

A7CC/A7CD is a new Latin case pair for s with diagonal stroke. The
proposal is silent about collation for these, so I'm just intercalating
them with a primary weight difference after A7A9/A7AA s with short
stroke overlay.

Generate allkeys.txt and verify that A7CC/A7CD show up properly weighted
as a case pair following A7A9/A7AA in primary order.

2. New Cyrillic case pair

1C89/1C8A is a new Cyrillic case pair for tje. The proposal is silent
about collation for these, but states that this is a Khanty letter for
[t'], i.e. [tʲ]. I'm intercalating it between Cyrillic letter twe and
Cyrillic letter Komi tje, with a primary weight distinction.

Generate allkeys.txt and verify that 1C89/1C8A show up properly weighted
as a case pair following A68C/A68D in primary order.

3. Archaic letter SHRI for Kannada and Telugu

0C5C;TELUGU ARCHAIC SHRII;Lo;;;;;;
0CDC;KANNADA ARCHAIC SHRII;Lo;;;;;;

The proposal claims that these can be collated as equivalent to the word SHRII,
i.e., the spelled out sequences. I'm adding the relevant sequences,
i.e. SHA-virama-RA-II, as a <sort> decomposition for each of these two:

0C5C;TELUGU ARCHAIC SHRII;Lo;<sort> 0C36 0C4D 0C30 0C40;;;;;
0CDC;KANNADA ARCHAIC SHRII;Lo;<sort> 0CB6 0CCD 0CB0 0CC0;;;;;

Generate allkeys.txt and verify that these two show up weighted correctly
by the decomposition.

4. Arabic Pegon letters

The proposal is silent about collation for these 4 additions.

0897;ARABIC PEPET;Mn;;;;;;

Because the pepet character was introduced to have a proper form for
what is currently commonly represented with a maddah, I give it a
<sort> equivalence to 0653 and intercalate it in unidata.txt right
after maddah. This is likely to provide better behavior for Pegon
material that may mix maddah and (in the future) 0897 pepet for this.
Revised entry in unidata.txt looks like this:

0897;ARABIC PEPET;Mn;<sort> 0653;;;;;

10EC2;ARABIC LETTER DAL WITH TWO DOTS VERTICALLY BELOW;Lo;;;;;;
10EC3;ARABIC LETTER TAH WITH TWO DOTS VERTICALLY BELOW;Lo;;;;;;
10EC4;ARABIC LETTER KAF WITH TWO DOTS VERTICALLY BELOW;Lo;;;;;;

Default collation for these nuktated Arabic letters is fairly arbitrary.
I intercalated 10EC2 just ahead of 0759. I intercalated 10EC3 just after
088C, and 10EC4 just after 08B4.

Generate allkeys.txt and verify that these pepet shows up weighted as
equivalent to maddah and the other 3 have primary weights in the expected
order.

5. Arabic combining alef overlay

The proposal is not explicit about collation for 10EFC, but implies that
it should be treated as equivalent to 0670 ARABIC LETTER SUPERSCRIPT ALEF.
I am simply equating it to 0670 in DUCET, similar to the way several
variant forms of maddah are equated to 0653.

Generate allkeys.txt and verify that 10EFC is weighted identically to
0670.

11 more down, 1049 to go.

Archive this delta 5:

unidata-16.0.0d5.txt (1557279 bytes, 10/07/2023)