UCA 16.0 delta 6
From Ken:

O.k., a new day. Time to start in on the meat of the matter -- the whole script
additions in primary order. I start with the unicameral script additions.
Those are a bit simpler than any new bicameral scripts (Garay for
16.0). (See above in the Delta 3 discussion for the detailed rationale for
the placement of the new scripts -- I won't replicate that discussion here,
but just proceed, based on the placement decisions already made.)

1. Kirat Rai

I start with this one, even though it has some complexities, because it
is fresh in our minds from the extended discussion of the unicodetools PR.
First move the entire range of Kirat Rai alphabetic records (in code point
order) into unidata-16.0.0d6.txt, starting after Tangsa. This includes all
the consonants, all the vowels signs, which in Kirat Rai are actually
full standalone letters, and the two killers:

16D40;KIRAT RAI SIGN ANUSVARA;Lm;;;;;;
...
16D6C;KIRAT RAI SIGN SAAT;Lm;;;;;;

I omit the three punctuation marks for now. Those end up elsewhere in
DUCET, and it is more efficient to deal with all the punctuation additions
in a separate delta later on, after the alphabetic runs have been
established.

As noted in discussion, the ANUSVARA, TONPI (bindu), and VISARGA for Kirat
Rai are encoded as standalone modifier letters, rather than as combining
marks, and we've already decided not to try equating them to the various
combining candrabindu, visarga, etc., that use generic weights shared
with the Devanagari archetypes of the combining marks. The proposal suggests
that they be given primary distinctions and be left in code point order.
It might be better to move them to the end of the list of consonants, but
without further rationale provided, the simplest solution here is to simply
do as the proposal suggests. Indeed, the proposal states that "This sort
order with anusvara, tonpi, and visarga sorting first has been approved
by AKRS." So we just leave it that way for DUCET.

The next complication results from canonical equivalences for three
Kirat Rai vowels: AI, O, AU. In cases like this, the way to get the sifter
to introduce contractions is to surround the records in question with
the CONTRACTION pragma, to wit:

CONTRACTION

16D68;KIRAT RAI VOWEL SIGN AI;Lo;16D67 16D67;;;;;
16D69;KIRAT RAI VOWEL SIGN O;Lo;16D63 16D67;;;;;
16D6A;KIRAT RAI VOWEL SIGN AU;Lo;16D69 16D67;;;;;

DEFAULT

In cases like this, as for Tamil, Kannada, etc., etc., I also drop
a comment into unidata.txt with a somewhat redundant explanation,
to remind everybody what is going on here.

The effect of the CONTRACTION pragma is to tell the sifter that for
the range of entries where it is in effect, the sifter is to go ahead and
assign a primary weight to the code point and *also* generate a
contraction entry from the decomposition, giving it the same weight
as the atomic character code point. In the absence of the CONTRACTION
pragma, such an entry is instead just entered into allkeys.txt with
the sequence of weights from the decomposition, and does not have
its own primary weight.

But wait, there's more. Because of the strange encoding of Kirat Rai
vowel signs, we have a canonical closure problem for 16D6A. The
full decomposition for 16D6A is <16D63, 16D67, 16D67>, and we need
to weight that sequence with the same primary weight via contraction.
Fortunately, this is not the first time this problem has been encountered
for the sifter. A similar problem of canonical closure for a recursive
canonical decomposition occurs for 0CCB in Kannada and 0DDD in Sinhala.
The mechanism baked in to the sifter to deal with this is a "secondary
decomposition", which can be added to the input entry in the decomposition
field. As a first step for handling 16D6A, I'm putting the following
entry into unidata-16.0.0d6.txt:

16D6A;KIRAT RAI VOWEL SIGN AU;Lo;16D69 16D67, 16D63 16D67 16D67;;;;;

The comma delimitation in the decomposition allows for adding a secondary
decomposition. When enclosed within the CONTRACTION pragma, this generates
a second contraction using the secondary decomposition information.

This *almost* solves the problem for Kirat Rai, as it was solved for
Kannada and Sinhala. Unfortunately, the way the Kirat Rai vowels work,
there is yet *another* sequence that is canonically equivalent:
<16D63, 16D68>. That sequence is equivalent to <16D63, 16D67, 16D67>,
so it also needs a contraction that is weighted with the same primary
weight. The sifter code currently treats the Kannada and Sinhala cases
essentially as the extent of the problem -- only a *single* secondary
contraction is allowed for in the code. The syntax is for this currently is

decompfield := decomp (, decomp)?

rather than

decompfield := decomp (, decomp)*

because the code was written to just look for the comma and then process
a *single* secondary decomposition value, rather than being written to
expect and process an indefinite *list* of secondary decompositions.

To fix this for Kirat Rai (and future-proof against any similar cases
in the future), I'm going to have to do some considerable refactoring
of the relevant decomposition handling code in the sifter, which is
a tricky and sensitive part of the code. For now
I am just going to postpone that code work until all the rest of the 16.0 input
for UCA has been taken care of.

But in the meantime, while adding the first relevant secondary decomposition
for 16D6A to unidata.txt, I also reversed the order of the secondary
decomposition for the 0CCB and 0DDD entries in unidata.txt. Now the field with
the secondary decomposition better matches what the code states, assuming
the *first* entry is the formal canonical decomposition string from
the UCD (which is then recursively decomposed internal to the sifter
processing), followed by a secondary decomposition, which in the Kannada
and Sinhala cases is the full decomposition. The impact on the output
in allkeys.txt is just to invert the order of two contraction lines,
but it does not affect any of the weighting per se. The other side effect
is that the output log will stop warning about encountering a "Non-binary"
canonical decomposition for 0CCB and 0DDD in the recursive decomposition.

Generate allkeys.txt and verify that Kirat Rai weights are as expected,
with special attention to the results for 16D68, 16D69, and 16D6A.
Also examine the impact of the secondary decomposition change for
0CCB and 0DDD.

O.k., this discussion for Kirat Rai and its implications is hairy enough
that I'm going to make this its own delta, without introducing more
scripts into this set of changes.

45 more down, 1004 to go.

Archive this delta 6:

unidata-16.0.0d6.txt (1559396 bytes, 10/08/2023)