UCA 16.0 delta 8
From Ken:

1. Garay

I've saved Garay for a separate delta, because it is the only new bicameral
script in the bunch, and that introduces intercalation complications for it
in unidata.txt. It also has a sukun, a gemination mark, and a reduplication mark.
The proposal implies it needs a syllabic ordering, rather than a simpler ordering.
It also has a couple variant letters, which need special handling.

Garay goes into unidata.txt after Medefaidrin, another West African bicameral
script and before Adlam, another West African bicameral script.

The first step is to move all the capital and small letters into unidata.txt.
Then I rearrange them into case pairs, in the same manner as for
Medefaidrin and Adlam. Note that this rearrangement is not strictly necessary
to get weight assignments correct for the case pairs, but it is better to
keep maintaining new bicameral scripts in the same way as for existing
ones already in unidata.txt. This consistency helps in understanding
what is going on.

Next, deal with OLD KA (10D64/10D84) and OLD NA (10D65/10D85). These are claimed
to just variant forms of KA and NA, respectively, and are claimed explicitly
to sort equal to them. I merge them into the case pair bundles for KA and NA,
with a <sort> decomposition to make them sort similarly to KA and NA, but
with a case-distinguished tertiary weight difference.

It is completely unclear how to weight the vowels. There is an extensive
discussion of collation in L2/22-048, but the main upshot of that seems to
be that collation is conceived of in terms of a syllabic grid, rather than
in terms of the actual string used to represent each node of the syllabic
grid. I default to putting the vowels in code point order ahead of all
the consonants. Any attempt to implement the actual syllabic ordering
would require an extensive tailoring. Just putting the vowels first, in
roughly the order specified for vowels in syllables, would seem to suffice
for the default ordering in DUCET.

The gemination mark (10D6A) will get a script-specific secondary, so I defer
that for now.

Regenerate allkeys.txt, and verify that the Garay weights are
as expected, including all the case pairs and the equivalents for the
two pairs of variant letters.

51 more down, 720 to go.

Archive this delta 8:

unidata-16.0.0d8.txt (1572176 bytes, 10/08/2023)