UCA 16.0 delta 1
From Ken:

Simply copy unidata-15.1.0d5.txt (7/28/2023 revision used for 15.1.0 release)
to unidata-16.0.0d1.txt and update internal date, version, etc.

Run existing 15.1.0 sifter executable with the 16.0.0 updated library for
properties, case mapping, etc., to process unidata-16.0.0d1.txt.
Verify that the output allkeys.txt is identical to the allkeys.txt
released for 15.1.0, except for the generated date header.

Current initial state archived as:

unidata-16.0.0d1.txt (1550494 bytes, 10/06/2023)

Process the diff between the released 15.1.0 UnicodeData.txt
(UnicodeData-15.1.0d3.txt) and the current latest draft of the
16.0.0 UnicodeData.txt (UnicodeData-16.0.0d7.txt). Clean this up
to just a list of all the 1177 new UnicodeData.txt records for
16.0. Run the results through a small transducer that snips out
the fields not used for the unidata.txt input for the sifter.
(This is just a simple utiity I have had for years -- it would
be easy to replicate in Perl or Python, as needed.) The result
is archived as:

uc151to160add.txt (49212 bytes, 10/07/2023)

That is the source of the data fields to paste into the evolving
draft of unidata.txt for 16.0. I do it this way because the
bookkeeping is all automatic. I need to find the right place in unidata.txt
for all 1177 lines, and when the input from uc151to160add.txt has
dwindled down to 0 lines left to transfer, I know I'm completely
done.