UCA 16.0 delta 12
From Ken:

I've done the code refactoring that enables the processing of more than
one secondary decomposition in the input file, so we could get the full
set of contractions working for the Kirat Rai vowel au.

There was a very
substantial refactoring of the code related to processContractions().
Instead of the prior one-off branch that checked for a secondary
decomposition and handled it specially, the code now assumes that a
decomposition in the input file is a comma-separated list which usually
defaults to a single entry, if present. However, rather than refactoring
the code so that it could handle indefinite lists of decompositions for
contractions, I just have it work now with a static array of up to four
decompositions. For over a decade, we've been able to get by with two.
Kirat Rai forces us up to three. I expect it will take awhile to find a
situation that requires four -- but if we do eventually need to go
there, the code will continue to just work with no further updates.
Doing it this way avoided even *more* extensive changes to the code to
build up and tear down dynamic lists for this extremely edgy edge case.