UCA 16.0: Moving non-decimal numerics after digits
For PAG issue 101 "re-align the DUCET & CLDR: order of groups below letters"

From Ken:

UCA 16.0 delta 17

This implements the move of the range of non-decimal numerics, so they
get primary weights *after* 0..9 and are no longer marked as variables.

The change to unidata.txt is diffable, although it involves a large
change: 570 lines of input for these numeric entries were moved down
from before the extenders to between the entries for DIGIT NINE (and
others numerically equivalent to 9) and LATIN SMALL LETTER A. And then
there are a few comment lines of explanation added.

The change to allkeys.txt is simply describable, but not really diffable
unless you ignore the primary weight assignments. The numerics are no
longer variables, but now have primary weights in the range 2187..237F,
so sort after DIGIT NINE (with primary weight of 2186), but ahead of
LATIN SMALL LETTER A. Primary weights from LATIN SMALL LETTER A onward
were unaffected, but the move of the numerics shifted all the primary
weights for extenders, currency signs, and digits. The size of the
generated file is identical to the previous one, which is a good sign.
The number of primary weights is also identical, as expected. The first
non-variable is still U+02D0 MODIFIER LETTER TRIANGULAR COLON, as
expected, but its primary weight is 212A, instead of 2323. I assume your
code automatically adjusts to identify the weight of the first non-variable.

I also regenerated decomps.txt. It isn't impacted by the numerics
rearrangement, but it does pick up the additional synthetic
decomposition added for the Tulu-Tigalari looped virama.

You will also need to pick up a small change to the sifter source code
in order to be able to replicate this output: sifter/unisift.c

The change is very small -- I simply had to comment out two lines in the
branch in the main sift dealing with numerics which set the identified
characters to variables. The rest just all falls out automatically given
the change in the input file.